requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   google-genai
    #   httpx
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via requests
cryptography==46.0.7
    # via google-auth
distro==1.9.0
    # via google-genai
google-auth==2.49.2
    # via google-genai
google-genai==1.73.1
    # via tjm-project (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
mouseinfo==0.1.3
    # via pyautogui
mss==10.1.0
    # via tjm-project (pyproject.toml)
pillow==12.2.0
    # via
    #   tjm-project (pyproject.toml)
    #   pyscreeze
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pyautogui==0.9.54
    # via tjm-project (pyproject.toml)
pycparser==3.0
    # via cffi
pydantic==2.13.1
    # via google-genai
pydantic-core==2.46.1
    # via pydantic
pygetwindow==0.0.9
    # via
    #   tjm-project (pyproject.toml)
    #   pyautogui
pymsgbox==2.0.1
    # via pyautogui
pyperclip==1.11.0
    # via
    #   tjm-project (pyproject.toml)
    #   mouseinfo
pyrect==0.2.0
    # via pygetwindow
pyscreeze==1.0.1
    # via pyautogui
python-dotenv==1.2.2
    # via tjm-project (pyproject.toml)
pytweening==1.2.0
    # via pyautogui
requests==2.33.1
    # via
    #   tjm-project (pyproject.toml)
    #   google-auth
    #   google-genai
sniffio==1.3.1
    # via google-genai
tenacity==9.1.4
    # via google-genai
typing-extensions==4.15.0
    # via
    #   anyio
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.6.3
    # via requests
websockets==16.0
    # via google-genai
